+++ api/src/main/java/org/grnet/cat/api/endpoints/AssessmentsV2Endpoint.java
@@ -810,6 +810,50 @@ public class AssessmentsV2Endpoint {
             this.content = content;
         }
     }
+    @Tag(name = "Assessment")
+    @Operation(
+            summary = "Get list of public assessment objects created / used by a specific user by type and actor.",
+            description = "This endpoint is public and any unauthenticated user can retrieve published assessment objects categorized by motivation and actor, created by all users." +
+                    "By default, the first page of 10 public assessment objects will be returned. You can tune the default values by using the query parameters page and size.")
+    @APIResponse(
+            responseCode = "200",
+            description = "List of public assessment objects.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = AssessmentsEndpoint.PageableObjects.class)))
+    @APIResponse(
+            responseCode = "400",
+            description = "Bad Request",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "404",
+            description = "Entity Not Found.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "500",
+            description = "Internal Server Error.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @GET
+    @Path("/public-objects/by-actor/{actor-id}")
+    @Produces(MediaType.APPLICATION_JSON)
+    public Response objectsByActor(       @PathParam("actor-id") @Valid @NotFoundEntity(repository = RegistryActorRepository.class, message = "There is no Actor with the following id:") String actorId,
+                                          @Parameter(name = "page", in = QUERY,
+                                                  description = "Indicates the page number. Page number must be >= 1.") @DefaultValue("1") @Min(value = 1, message = "Page number must be >= 1.") @QueryParam("page") int page,
+                                          @Parameter(name = "size", in = QUERY,
+                                                  description = "The page size.") @DefaultValue("10") @Min(value = 1, message = "Page size must be between 1 and 100.")
+                                          @Max(value = 100, message = "Page size must be between 1 and 100.") @QueryParam("size") int size,
+                                          @Context UriInfo uriInfo) {
+
+        var assessments = assessmentService.getPublishedAssessmentObjectsByActorAndPage(page - 1, size, actorId,uriInfo);
+
+        return Response.ok().entity(assessments).build();
+    }
 
 
 }